backend/requirements.txt
# Python FastAPI Backend Requirements
# Install with: pip install -r requirements.txt

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# CORS Support
python-multipart==0.0.6

# Quantum Simulation (NumPy for calculations)
numpy==1.26.3

# Cryptography
pycryptodome==3.20.0

# HTTP Client (for testing)
httpx==0.26.0

# Development
python-dotenv==1.0.0
